Moved CGCreature to a new file
